requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
drf-yasg>=1.21.7
django-redis>=5.3.0
psycopg2-binary>=2.9.6
redis>=4.5.5
PyJWT>=2.8.0
python-decouple>=3.8
gunicorn>=21.2.0